pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pipelearn"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Tilak1114", email = "tilaksharma1114@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "numpy==1.26.4",
  "Pillow==10.4.0",
  "setuptools==68.2.2",
  "torch==2.2.2",
  "transformers==4.44.2",
  "pydantic==2.7.4",
  "pytest==8.2.2",
  "embdata"
]

[project.urls]
Documentation = "https://github.com/Tilak1114/pipelearn#readme"
Issues = "https://github.com/Tilak1114/pipelearn/issues"
Source = "https://github.com/Tilak1114/pipelearn"

[tool.hatch.version]
path = "pipelearn/__about__.py"

[tool.hatch.envs.default]
path = ".venv"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:pipelearn tests}"

[tool.coverage.run]
source_pkgs = ["pipelearn", "tests"]
branch = true
parallel = true
omit = [
  "pipelearn/__about__.py",
]

[tool.coverage.paths]
pipelearn = ["pipelearn", "*/pipelearn/pipelearn"]
tests = ["tests", "*/pipelearn/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
